requirements.txt
tensorflow==2.15.0
xgboost
joblib
numpy
Pillow
